Separate out expected.hh
