+++ pdns/recursordist/expected.hh
@@ -1,0 +1,1 @@
+../expected.hh
